Refactors scoreboard styles
