+++ src/components/scoreboard/Scoreboard.module.css
@@ -1,19 +1,32 @@
-@value blackBoard: #282828;
 @value chalkWhite: #fefeff;
 
+.body,
 .cell,
+.foot,
+.head,
 .table {
   display: grid;
 }
 
+.body,
+.foot,
+.head {
+  column-gap: 0.1rem;
+  grid-template-columns: repeat(2, 1fr);
+}
+
+.body {
+  overflow: auto;
+}
+
 .cell {
   color: chalkWhite;
+  column-gap: 0.25rem;
   font-size: 1.25rem;
-  grid-column-gap: 0.25rem;
   grid-template-areas: ". content emoji";
   grid-template-columns: repeat(3, 1fr);
-  text-align: center;
   padding: 0.5rem;
+  text-align: center;
   text-transform: uppercase;
 }
 
@@ -22,7 +35,7 @@
 }
 
 .column {
-  background-color: blackBoard;
+  background-color: #282828;
 }
 
 .content {
@@ -31,6 +44,19 @@
 
 .emoji {
   grid-area: emoji;
+}
+
+.foot,
+.head {
+  font-weight: bold;
+}
+
+.foot {
+  border-top: 0.2rem solid chalkWhite;
+}
+
+.head {
+  border-bottom: 0.2rem solid chalkWhite;
 }
 
 .table {
