--- src/components/scoreboard/body/Body.module.css
@@ -1,6 +1,0 @@
-.body {
-  column-gap: 0.1rem;
-  display: grid;
-  grid-template-columns: repeat(2, 1fr);
-  overflow: auto;
-}
