+++ src/components/scoreboard/foot/Foot.tsx
@@ -1,14 +1,13 @@
 import React, { FC } from "react";
 import { Props } from "./types";
-import scoreboardStyles from "../Scoreboard.module.css";
-import styles from "./Foot.module.css";
+import styles from "../Scoreboard.module.css";
 
 export const Foot: FC<Props> = ({ footCellRef, footRef, score, teams }) => (
   <div className={styles.foot} ref={footRef}>
     {teams.map(team => (
-      <div className={scoreboardStyles.column} key={team}>
-        <div className={scoreboardStyles.cell} ref={footCellRef}>
-          <div className={scoreboardStyles.content}>{score[team]}</div>
+      <div className={styles.column} key={team}>
+        <div className={styles.cell} ref={footCellRef}>
+          <div className={styles.content}>{score[team]}</div>
         </div>
       </div>
     ))}
