+++ src/components/scoreboard/head/Head.tsx
@@ -3,8 +3,7 @@ import { Props } from "./types";
 import { Team } from "shared/types";
 import classNames from "classnames";
 import { last } from "lodash";
-import scoreboardStyles from "../Scoreboard.module.css";
-import styles from "./Head.module.css";
+import styles from "../Scoreboard.module.css";
 
 export const Head: FC<Props> = ({ headRef, score, teams }) => {
   const [leader, setLeader] = useState<Team | null>(null);
@@ -18,17 +17,17 @@ export const Head: FC<Props> = ({ headRef, score, teams }) => {
   return (
     <div className={styles.head} ref={headRef}>
       {teams.map(team => (
-        <div className={scoreboardStyles.column} key={team}>
+        <div className={styles.column} key={team}>
           <div
-            className={classNames(scoreboardStyles.cell, {
-              [scoreboardStyles.reverse]: last(teams) === team
+            className={classNames(styles.cell, {
+              [styles.reverse]: last(teams) === team
             })}
           >
-            <div className={scoreboardStyles.content}>{team}</div>
+            <div className={styles.content}>{team}</div>
             {leader === team && (
               <span
                 aria-label="Chequered Flag"
-                className={scoreboardStyles.emoji}
+                className={styles.emoji}
                 role="img"
               >
                 🏁
